requirements.txt
pdfplumber>=0.10.0
fpdf2>=2.7.0
matplotlib>=3.8.0
google-genai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=10.0.0
feedparser>=6.0.0
streamlit>=1.30.0
plotly>=5.18.0
